Moved UIImageView download to separate extension
