+++ Sources/PrimerSDK/Classes/Extensions & Utilities/UIImageView+Extensions.swift
@@ -1,0 +1,35 @@
+//
+//  UIImageView+Extensions.swift
+//  PrimerSDK
+//
+//  Copyright © 2022 Primer API ltd. All rights reserved.
+//
+
+#if canImport(UIKit)
+
+import UIKit
+
+extension UIImageView {
+    
+    func downloaded(from url: URL, contentMode mode: ContentMode = .scaleAspectFit) {
+        contentMode = mode
+        URLSession.shared.dataTask(with: url) { data, response, error in
+            guard
+                let httpURLResponse = response as? HTTPURLResponse, httpURLResponse.statusCode == 200,
+                let mimeType = response?.mimeType, mimeType.hasPrefix("image"),
+                let data = data, error == nil,
+                let image = UIImage(data: data)
+                else { return }
+            DispatchQueue.main.async() { [weak self] in
+                self?.image = image
+            }
+        }.resume()
+    }
+    
+    func downloaded(from link: String, contentMode mode: ContentMode = .scaleAspectFit) {
+        guard let url = URL(string: link) else { return }
+        downloaded(from: url, contentMode: mode)
+    }
+}
+
+#endif
